Add default lao channel
