+++ fe1-web/src/features/lao/functions/index.ts
@@ -1,3 +1,4 @@
 export * from './lao';
+export * from './network';
 export * from './TestConnection';
 export { requestCreateLao } from '../network';
